backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-Cors==4.0.0 
gunicorn==22.0.0
PyJWT==2.8.0
requests==2.31.0 
SQLAlchemy==2.0.29
PyMySQL==1.1.0
